refactor simulation files
