+++ optosim/simulation/__init__.py
@@ -1,0 +1,14 @@
+from . import utils
+from .utils import *
+
+from . import ptfe_scatter
+from .ptfe_scatter import *
+
+from . import optical_photon
+from .optical_photon import *
+
+from . import generator
+from .generator import *
+
+from . import event_reader
+from .event_reader import *
