+++ optosim/simulation/utils.py
